+++ src/lib/components/home-page/news/NewsSkeletonShell.svelte
@@ -1,16 +1,16 @@
 <script lang="ts">
+  import SkeletonShell from "$lib/components/SkeletonShell.svelte";
+
   export let horizontalLine = true;
 </script>
 
 <section>
-  <div class="title" />
-  <div class="flex">
-    <div class="author">
-      <span class="avatar" />
-      <span class="author-name" />
-    </div>
-    <div class="date" />
-  </div>
+  <SkeletonShell width="100%" height="64" primaryColor="#EBECEFAA" secondaryColor="#F5F5F7EE">
+    <rect width="70%" height="1.3em" x="0" y="0" rx="4" ry="4" />
+    <circle r="16" cx="16" cy="3em" />
+    <rect width="40%" height="1em" x="38" y="2.5em" rx="4" ry="4" />
+    <rect width="20%" height="0.8em" x="80%" y="2.66em" rx="4" ry="4" />
+  </SkeletonShell>
   {#if horizontalLine}
     <div class="horizontal-line" />
   {/if}
@@ -26,43 +26,5 @@
     position: relative;
     padding: 0.75em $news-bleed-padding 0.5em;
     margin: 0 #{$news-bleed-padding * -1} 0 #{$news-bleed-padding * -1};
-  }
-  .title {
-    height: 1.3em;
-    width: var(--width, 40%);
-    border-radius: 2px;
-    background-color: $color;
-    margin-bottom: 12px;
-  }
-
-  .flex {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    .author {
-      display: flex;
-      align-items: center;
-      .avatar {
-        $size: 32px;
-        width: $size;
-        height: $size;
-        border-radius: 50%;
-        background-color: $color;
-        margin-right: 6px;
-      }
-      .author-name {
-        height: 1.1em;
-        width: 72px;
-        border-radius: 2px;
-        background-color: $color;
-      }
-    }
-
-    .date {
-      height: 0.75em;
-      width: 100px;
-      border-radius: 2px;
-      background-color: $color;
-    }
   }
 </style>